+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * DIM is generally just a little bit snappier, especially when scrolling.
 * Clicking the icon to open DIM will now switch to an active DIM tab if it's already running.
 * Bungie.net will open in a new tab as a convenience for expired cookies.
+* Items in the Postmaster are sorted by the order you got them, so you know what'll get bumped when your postmaster is full.
 
 # 3.7.4